Tidy up the tensorboard.main API
This change will make it easier to write plugins with custom builds of
TensorBoard. The plugin author will no longer be required to explicitly list
all the first-party plugins, as they can now be inherited via the
`tensorboard.default` module.

The `tensorboard.program` module has also been introduced. This module gives a
permanent home to all the APIs that crept into `tensorboard.main` over the
years. After this change, the `tensorboard.main` module is nothing more than
glue between the `program` and `default` modules.

The special-case code in main.py that the debugger plugin was using, has now
been relocated into the debugger plugin's directory.

This change is backwards-compatible. Delegates have been added for relocated
functions that will log a deprecation warning if used.
