+++ tensorboard/plugins/core/BUILD
@@ -3,6 +3,9 @@
 
 package(default_visibility = ["//tensorboard:internal"])
 
+load("//tensorboard/defs:web.bzl", "ts_web_library")
+load("//tensorboard/defs:zipper.bzl", "tensorboard_zip_file")
+
 licenses(["notice"])  # Apache 2.0
 
 exports_files(["LICENSE"])
@@ -25,6 +28,7 @@ py_test(
     name = "core_plugin_test",
     size = "small",
     srcs = ["core_plugin_test.py"],
+    data = ["test_webfiles.zip"],
     srcs_version = "PY2AND3",
     deps = [
         ":core_plugin",
@@ -35,4 +39,17 @@ py_test(
         "@org_pocoo_werkzeug",
         "@org_pythonhosted_six",
     ],
+)
+
+ts_web_library(
+    name = "test_assets",
+    testonly = 1,
+    srcs = ["index.html"],
+    path = "/",
+)
+
+tensorboard_zip_file(
+    name = "test_webfiles",
+    testonly = 1,
+    deps = [":test_assets"],
 )
