+++ tensorboard/plugins/core/index.html
@@ -1,0 +1,2 @@
+<!doctype html>
+<p>This is a fake TensorBoard index page.
